requirements/development.txt
-r test.txt
django-extensions
werkzeug
django-debug-toolbar
bpython
django-template-debug
